Update Transcoder.mediafile.Metadata on SetInputPath
Adding the Metadata for files when using SetInputPath allows for
progress tracking when using a inputPath and a outputPipe. The example
below should be able to track progress percentage but can't before this
commit due to the lack of the duration in the Metadata.
```
err := trans.InitializeEmptyTranscoder()
// handle errors ...
err = trans.SetInputPath(input)
// handle errors ...
r, err := trans.CreateOutputPipe("mp4")
// handle errors ...
```

It should also fix the potential issue of having the MetaData for
another file if SetInputPath is called after initialising the transcoder
with `Transcoder.Initialize(inputPath, outputPath)`.